requirements.txt
liquer-framework>=0.6.5
liquer-gui
flask>=1.0.2
numpy>=1.15.4
pandas>=0.23.4
pytz>=2018.7
xlsxwriter>=1.1.2
pyparsing>=2.4.0
requests>=1.25.8
pyyaml
pyinstaller
openpyxl
matplotlib
